+++ HISTORY.md
@@ -3,3 +3,12 @@
 ## 0.1.0 (2019-10-04)
 
 * Package skeleton
+
+## 0.2.0 (2019-10-08)
+
+* Feature Importance Extraction from different interpretability models:
+  * Direct feature importances
+  * Global surrogate models
+  * LIME
+  * SHAP
+* Updated documentation served [here](https://tommartensen.de/tic).